feat: add helpers for deferred input/output etc. item access
