+++ snakemake/ioutils/lookup.py
@@ -221,115 +221,3 @@ def lookup(
         return DpathWildcardHandler(do_dpath, **namespace).handle(dpath)
     else:
         raise error("Must provide either a query or dpath parameter.")
-
-
-def evaluate(expr: str):
-    """Evaluate a python expression while replacing any wildcards given as
-    {wildcardname} with the wildcard value represented as a string."""
-
-    def inner(wildcards):
-        return eval(
-            expr.format(**{w: repr(v) for w, v in wildcards.items()}), globals()
-        )
-
-    return inner
-
-
-def branch(
-    condition: Union[Callable, bool],
-    then: Optional[Union[str, list[str], Callable]] = None,
-    otherwise: Optional[Union[str, list[str], Callable]] = None,
-    cases: Optional[Mapping] = None,
-):
-    """Branch based on a condition that is provided as a function pointer (i.e. a Callable)
-    or a value.
-
-    If then and optionally otherwise are specified, do the following:
-    If the condition is (or evaluates to) True, return the value
-    of the then parameter. Otherwise, return the value of the otherwise parameter.
-
-    If cases is specified, do the following:
-    Retrieve the value of the cases mapping using the return value of the condition
-    (if it is a function), or the condition value itself as a key.
-
-    The given condition function has to take wildcards as its only parameter.
-    Similarly, then, otherwise and the values of the cases mapping can be such functions.
-
-    If any such function is given to any of those arguments, this function returns a derived
-    input function that will be evaluated once the wildcards are known.
-    """
-
-    def convert_none(value):
-        return value or []
-
-    def handle_callable(value, wildcards):
-        if isinstance(value, Callable):
-            return convert_none(value(wildcards))
-        else:
-            return convert_none(value)
-
-    def do_branch_then_otherwise(wildcards):
-        if handle_callable(condition, wildcards):
-            return handle_callable(then, wildcards)
-        else:
-            return handle_callable(otherwise, wildcards)
-
-    def do_branch_cases(wildcards):
-        res = handle_callable(condition, wildcards)
-        selected_case = cases[res]
-        return handle_callable(selected_case, wildcards)
-
-    do_branch = do_branch_then_otherwise
-    if cases is not None:
-        if otherwise is not None or then is not None:
-            raise ValueError("Cannot use cases together with then or otherwise.")
-        do_branch = do_branch_cases
-
-    if any(isinstance(value, Callable) for value in (condition, then, otherwise)):
-
-        def inner(wildcards):
-            return do_branch(wildcards)
-
-        return inner
-    else:
-        return do_branch(None)
-
-
-# Alias for expand that provides a more intuitive name for the use case of
-# collecting files from previous jobs.
-collect = snakemake.io.expand
-
-
-def exists(path):
-    """Return True if the given file or directory exists.
-
-    This function considers any storage arguments given to Snakemake.
-    """
-    func_context = inspect.currentframe().f_back.f_locals
-    func_context_global = inspect.currentframe().f_back.f_globals
-
-    workflow = func_context.get("workflow") or func_context_global.get("workflow")
-
-    if workflow is None:
-        raise WorkflowError(
-            "The exists function can only be used within a Snakemake workflow "
-            "(the global variable 'workflow' has to be present)."
-        )
-
-    path = workflow.modifier.path_modifier.apply_default_storage(path)
-    if snakemake.io.is_flagged(path, "storage_object"):
-        return async_run(path.flags["storage_object"].managed_exists())
-    else:
-        return os.path.exists(path)
-
-
-def register_in_globals(_globals):
-    _globals.update(
-        {
-            "lookup": lookup,
-            "evaluate": evaluate,
-            "branch": branch,
-            "collect": collect,
-            "exists": exists,
-        }
-    )
